renamed TimestampNTZ to SnowflakeTimestampNTZAsUTC
